Optimized DomRiaOutputConverter tests
