+++ .pylintrc
@@ -1,2 +1,2 @@
 [MESSAGES CONTROL]
-disable=R0201, C0413, R0903, W1113, W0613, W0231, R0401, C0415, W0603, R0914, R1702, R0912
+disable=R0201, C0413, R0903, W1113, W0613, W0231, R0401, C0415, W0603, R0914, R1702, R0912, W0621,
