+++ tests/conftest.py
@@ -2,4 +2,4 @@
 Config module for pytest
 """
 
-pytest_plugins = ["converter_fixures"]
+pytest_plugins = ["converter_fixtures"]
